feat (apps) Custom Model/Manager for users
Previous: 2023-10-02 00.00.00.002
Changelog: 2023-10-02 00.00.00.003
Replica: https://github.com/ipoetdev/dash-and-do
Replicat: Branch: epic-kore
Intent: Setting up Admin with custom user and manager
Goal: To add a user to the database via Admin or replciate bug in admin

AI Notes
Add custom user model and registration system

Added a custom user model to handle user authentication and registration. It includes changes to the settings file, where the default User model was replaced with the new one; and creation of a manager for the model that defines methods required to create instances of 'User'. This update is necessary to have more control over user information and authentication process. This also includes creation of a new app for managing user activities. The Django-allauth package used to handle user registration and auth related tasks. These modifications were created on the basis of requirements for managing user registration and profile management in a more customized and flexible manner.